feat(utils): add slugify utility function
The slugify utility function is added to the utils module. This function takes a string as input and returns a slugified version of the string. It uses the slugify library to perform the slugification. The function also accepts optional options to customize the slugification process, such as specifying a replacement character, removing certain characters, converting to lowercase, and more. The default options are set to use the English locale and convert the slug to lowercase.