refactor(button): change shape property value from string to array to allow for multiple shape options
The shape property in the buttonVariants object in Button.tsx has been changed from a string to an array. This allows for multiple shape options to be specified for each size variant of the button. The shape options now include both "icon" and "circle". This change improves the flexibility and customization options for the button component.