ui update: body default classes, image, card, drawer, descriptions, sidebar, checkbox (#115)
* feat: implement tanstack router hooks and configuration

Add custom hooks for routing, search parameters, and pathname 
management using TanStack Router. Create a package.json for 
module configuration, including build and lint scripts. 
Establish ESLint configuration for code quality and 
TypeScript settings for type safety. This enhances routing 
capabilities and streamlines development processes.

* style: update body styles for improved readability

Enhance the body styles by adding min-h-screen and font-sans 
to ensure a minimum height and a more consistent font. This 
improves overall readability and visual appeal while maintaining 
the existing background and text color settings.

* fix(card): ensure CardTitle uses full width in flex layout

Update CardTitle component to include "flex-1" class for proper width 
allocation. This change addresses the issue of the title not utilizing 
the full available width when rendered in a flex container.

* feat(checkbox): add loading state to Checkbox component

Introduce a loading prop to the Checkbox component to display a 
loading icon when the checkbox is in a loading state. This change 
enhances user experience by providing visual feedback during 
asynchronous operations. Adjust the component's layout to ensure 
proper alignment when loading.

* feat(descriptions): enhance className handling for columns

Adds support for custom class names in the Descriptions component. 
This change allows for more flexible styling by incorporating 
`col.classNames?.label` and `col.classNames?.value` into the 
`th` and `td` elements, respectively. This improves the 
component's adaptability to different design requirements.

* feat(drawer): switch sheet to vaul

Adds a new Sheet component with customizable content and variants. 
Refactors imports in the sidebar to streamline dependencies and 
improve organization. This enhances the modularity and usability 
of the UI components.

* feat(icons): add info-filled icon to the icon library

Adds the InfoFilled icon to the icon library by creating a new 
component in `info-filled.tsx` and exporting it from the index file. 
This enhances the icon collection and provides a new visual 
element for use in the application.

* feat(image): Image component with preview and fallback

Introduce a new Image component that supports loading states, error 
handling, and an optional preview feature. The component utilizes 
placeholders while the image loads and displays a fallback image 
in case of an error. This enhances the user experience by providing 
visual feedback during image loading and error scenarios.

* feat(sidebar): enhance sidebar item rendering and props

Adds an `itemRender` prop to customize item rendering in the 
Sidebar component. Updates the label rendering logic to support 
customization and modifies the items rendering to repeat items 
for demonstration purposes. This improves flexibility and 
extensibility of the Sidebar component.

* refactor: optimize SidebarContent rendering logic

Removes redundant item duplication in SidebarContent. 
This change improves performance and simplifies the 
rendering logic by ensuring that only the original 
items are passed to the render function.