ui update: sidebar itemRender, datepicker value (#124)
* feat(sidebar): enhance sidebar item rendering and props

Adds an `itemRender` prop to customize item rendering in the 
Sidebar component. Updates the label rendering logic to support 
customization and modifies the items rendering to repeat items 
for demonstration purposes. This improves flexibility and 
extensibility of the Sidebar component.

* refactor: optimize SidebarContent rendering logic

Removes redundant item duplication in SidebarContent. 
This change improves performance and simplifies the 
rendering logic by ensuring that only the original 
items are passed to the render function.

* refactor(date-range-picker): improve value handling and types

Update the `DateRangePicker` component to ensure that the `onChange` 
callback only triggers when both start and end dates are defined. 
Refine the type definitions for `value` and `onChange` to prevent 
undefined values, enhancing type safety and clarity. Adjust the 
default month and selected date handling to accommodate these 
changes, ensuring consistent behavior across the component.

* refactor(toast): update className order in Toaster component

Rearrange the className in the Toaster component to ensure 
consistency with Tailwind CSS conventions. This change improves 
readability and maintains adherence to best practices for 
class ordering.

* refactor(image): update Image component export structure

Change the export of the Image component to use a named export for 
better clarity and consistency. Also, export the ImageProps type 
to allow for easier type checking and usage in other components.

* chore: update package dependencies across projects

Update various package dependencies in the UI, ESLint, and 
prettier configurations to their latest versions. This includes 
updating libraries such as `sonner`, `swiper`, `tailwind-merge`, 
and several Radix UI components. These changes improve 
compatibility, enhance performance, and ensure access to the 
 features and bug.

* chore: run prettier