fix: build issue (#128)
* feat(invoice): add booking name to invoice booking table

Add a new column for "Booking Name" in the invoice booking table to 
enhance the information displayed to users. Update the database type 
to include the new field and adjust the table's visibility based on 
the booking count status. This improves user experience by providing 
more context for each booking.

* Uncommited changes

* refactor: status count (invoice, payroll)

* fix: build issue

* chore: update dependencies

* fix: go home not redirect to dashboard

* chore: remove dist_